refactor: 폴더명을 단수형으로 통일
