+++ src/domain/folder/component/setting-folder/folder-modal/FolderModal.tsx
@@ -1,7 +1,7 @@
 import Modal from "react-modal";
 import MemberSection from "./member-section/MemberSection";
-import { Folder } from "../../../layout/sidebar/SideBarType";
 import SubscriptionSection from "./sbscription-section/SubscriptionSection";
+import { Folder } from "../../../../layout/type/SideBarType";
 
 type Props = {
   isFolderModalOpen: boolean;
