+++ src/domain/folder/component/setting-folder/folder-modal/sbscription-section/SubscriptionSection.tsx
@@ -1,9 +1,9 @@
-import { API_PATH } from "../../../../../../constants/ApiPath";
-import { Blog, Folder } from "../../../../layout/sidebar/SideBarType";
+import { API_PATH } from "../../../../../../constant/ApiPath";
 import BlogBox from "./BlogBox";
 import { useState } from "react";
 import authAxios from "../../../../../../utill/ApiUtills";
 import { useFoldersStore } from "../../../../../../store/store";
+import { Blog, Folder } from "../../../../../layout/type/SideBarType";
 
 type Props = {
   folder?: Folder;
