+++ src/domain/layout/type/SideBarType.ts
