+++ src/domain/post/component/FolderPostListContent.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { Post } from "../type/PostType";
 import authAxios from "../../../utill/ApiUtills";
-import { API_PATH } from "../../../constants/ApiPath";
+import { API_PATH } from "../../../constant/ApiPath";
 import PostItemList from "./PostItemList";
 
 type Props = {
@@ -10,7 +10,7 @@ type Props = {
 
 export default function FolderPostListContent({folderId}: Props) {
 
-  let timeoutId: number;
+  let timeoutId: NodeJS.Timeout;
 
   const [page, setPage] = useState<number>(0);
   const [posts, setPosts] = useState<Post[]>([]);
